backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.1
firebase-admin==6.5.0
fastapi==0.111.0
fastapi-cli==0.0.4
fastapi-pagination==0.12.26
google-api-core==2.19.1
google-auth==2.31.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-secret-manager==2.20.1
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.5
packaging==24.1
pg8000==1.31.2
pluggy==1.5.0
psycopg2-binary==2.9.9
pydantic==2.8.0
pydantic_core==2.20.0
pydantic-settings==2.0.3
Pygments==2.18.0
pyrebase4==4.8.0
pytest==8.2.2
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
tomli==2.0.1
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
